hamster/requirements.txt
flask==3.0.3
pyjwt==2.7.0
bcrypt==4.1.2
cryptography==43.0.0
urllib3==2.0.7
gunicorn==23.0.0
elasticsearch      
elastic-apm[flask] 
python-json-logger 
